refactor(emails): fold _i18n folder into _lib/i18n.tsx single source of truth
Per PR review feedback — consolidate the email-internationalisation
surface into one file under emails/_lib/:
- Move emails/_i18n/index.ts content (EmailLocale, DEFAULT_LOCALE,
  loadMessages, createEmailTranslator, EmailMessages, EmailTranslator)
  into emails/_lib/i18n.tsx
- Move emails/_lib/intlRenderers.tsx content (renderStrong,
  renderSupportLink, createLinkRenderer) into emails/_lib/i18n.tsx
- Delete the now-empty emails/_i18n/ directory and the old
  intlRenderers.tsx file
- Update all import paths in email templates (~18 files)

The new emails/_lib/i18n.tsx is the single source of truth for
locale data, message-bundle loading, translator construction, and
shared t.rich() render functions.
