+++ sites/arolariu.ro/emails/_components/layoutTranslator.test.ts
@@ -1,6 +1,6 @@
 import {beforeEach, describe, expect, it, vi} from "vitest";
 
-import * as i18n from "../_i18n";
+import * as i18n from "../_lib/i18n";
 import {__resetLayoutTranslatorCache, getLayoutTranslator} from "./layoutTranslator";
 
 const FIXTURE = {email: {layout: {tagline: "Tag", buttonFallback: "BF"}}};
