+++ sites/arolariu.ro/emails/invoices/reminders/WeeklyUploadReminderEmail.tsx
@@ -22,8 +22,7 @@
 
 import {Text} from "react-email";
 import {BRAND, BulletList, EmailCard, EmailLayout, EmailParagraphStyles, MetricsGrid} from "../../_components";
-import {createEmailTranslator, DEFAULT_LOCALE, type EmailLocale, loadMessages} from "../../_i18n";
-import {renderSupportLink} from "../../_lib/intlRenderers";
+import {createEmailTranslator, DEFAULT_LOCALE, type EmailLocale, loadMessages, renderSupportLink} from "../../_lib/i18n";
 
 /**
  * Properties for the WeeklyUploadReminderEmail component.
